[PSL-401] Add support for Cezanne upgrade with activation height (#148)
- activate Cezanne upgrade (protocol 170009) by height
  used different activation parameters for mainnet, testnet and regtest
- changed min protocol version to 170008.
Peers with old version will still be able to connect until Cezanne activation height.
Starting from that height all nodes with old version will be banned.
- added nNetworkUpgradePeerPreferenceBlockPeriod parameter to consensus params -The period before a network upgrade activates, where connections to upgrading peers are preferred (in blocks).
- set different values for this parameter for different networks

+++ src/key.h
@@ -93,7 +93,7 @@ public:
     }
 
     //! Simple read-only vector-like interface.
-    unsigned int size() const noexcept { return (fValid ? keydata.size() : 0); }
+    unsigned int size() const noexcept { return (fValid ? static_cast<int>(keydata.size()) : 0); }
     const unsigned char* begin() const noexcept { return keydata.data(); }
     const unsigned char* end() const noexcept { return keydata.data() + size(); }
     const unsigned char* cbegin() const noexcept { return keydata.data(); }
